+++ OrderFlowTools/LadderNotesReader.cs
@@ -1,12 +1,8 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.IO;
-using System.Threading.Tasks;
 using NinjaTrader.NinjaScript.Indicators;
+using System;
 using System.Collections.Concurrent;
-using NinjaTrader.Cbi;
+using System.IO;
+using System.Linq;
 using System.Net;
 using System.Text.RegularExpressions;